fixed black camera while starting in water
Issue seemed to be caused by the layering of cameras doing post processing effects. By telling the base camera not to do post processing, the second overlay camera (which renders the item) can do all post processing. This seems to fix the issue, but post processing volumes will need to be changed. As it is now 1 camera rendering post processing and not 2, performance should also increase